build otimizado usando json

+++ scripts/mesclarOutros.js
@@ -1,10 +1,8 @@
-// scripts/mesclarOutros.js
 import fs from 'fs';
 import path from 'path';
 import fse from 'fs-extra';
 
-const tipo = process.argv[2]; // "produto", "mizuno", etc.
-
+const tipo = process.argv[2];
 if (!tipo) {
   console.error('❌ Você deve informar o tipo. Ex: node scripts/mesclarOutros.js produto');
   process.exit(1);
@@ -16,16 +14,25 @@ const diretorioFinal = path.join(process.cwd(), 'out');
 const diretoriosLote = fs
   .readdirSync(process.cwd())
   .filter(dir => regex.test(dir))
- // .filter(dir => /^out-[a-z0-9]+-lote-\d+$/i.test(dir)) // contempla todas iniciando com out-xxx-lote-xx
   .map(dir => path.join(process.cwd(), dir));
-
-// Garante que a pasta final esteja limpa
-/*if (fs.existsSync(diretorioFinal)) {
+/*
+// Limpa diretório final /out
+if (fs.existsSync(diretorioFinal)) {
   console.log('🧹 Limpando diretório final: /out');
   fs.rmSync(diretorioFinal, { recursive: true, force: true });
-}*/
-
-//fs.mkdirSync(diretorioFinal);
+}
+fs.mkdirSync(diretorioFinal);
+*/
+// Copia tudo da pasta public para out
+const publicDir = path.join(process.cwd(), 'public');
+if (fs.existsSync(publicDir)) {
+  fse.copySync(publicDir, diretorioFinal, {
+    overwrite: true,
+    dereference: true,
+    errorOnExist: false
+  });
+  console.log(`📦 Copiado conteúdo de public para /out`);
+}
 
 console.log(`📦 Mesclando ${diretoriosLote.length} lotes para /out...`);
 
@@ -36,7 +43,7 @@ for (const dir of diretoriosLote) {
     errorOnExist: false
   });
   console.log(`✅ Copiado: ${dir}`);
-  
+
   fs.rmSync(dir, { recursive: true, force: true });
   console.log(`🗑️  Apagado: ${dir}`);
 }
